Move handler declaration di container
