+++ app/Ownership/OwnershipRequestHandlerServiceProvider.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace CultuurNet\UDB3\Ownership;
 
 use CultuurNet\UDB3\Container\AbstractServiceProvider;
-use CultuurNet\UDB3\Http\Organizer\GetCreatorRequestHandler;
 use CultuurNet\UDB3\Http\Ownership\ApproveOwnershipRequestHandler;
 use CultuurNet\UDB3\Http\Ownership\DeleteOwnershipRequestHandler;
 use CultuurNet\UDB3\Http\Ownership\GetOwnershipRequestHandler;
@@ -26,7 +25,6 @@ final class OwnershipRequestHandlerServiceProvider extends AbstractServiceProvid
             OwnershipStatusGuard::class,
             RequestOwnershipRequestHandler::class,
             GetOwnershipRequestHandler::class,
-            GetCreatorRequestHandler::class,
             SearchOwnershipRequestHandler::class,
             ApproveOwnershipRequestHandler::class,
             RejectOwnershipRequestHandler::class,
@@ -63,16 +61,6 @@ final class OwnershipRequestHandlerServiceProvider extends AbstractServiceProvid
             GetOwnershipRequestHandler::class,
             fn () => new GetOwnershipRequestHandler(
                 $container->get(OwnershipServiceProvider::OWNERSHIP_JSONLD_REPOSITORY)
-            )
-        );
-
-        $container->addShared(
-            GetCreatorRequestHandler::class,
-            fn () => new GetCreatorRequestHandler(
-                $container->get('organizer_jsonld_repository'),
-                $container->get(UserIdentityResolver::class),
-                $container->get('organizer_permission_voter'),
-                $container->get(CurrentUser::class),
             )
         );
